refactor(parity): move the DupCoder citations into JSDoc, drop the non-Rails ensure-through guard from the proxy delegations, and split the trails-only DupCoder cases out of cache.test.ts
