+++ packages/activesupport/src/cache/memory-store.ts
@@ -157,9 +157,11 @@ export class MemoryStore extends Store implements CacheStore {
     return num;
   }
 
-  // Rails guards `prune` against re-entry with the `@pruning` flag
-  // (memory_store.rb:110-127): `cleanup` and the per-key `delete_entry` below can
-  // both re-enter through a write, and the inner call must not prune again.
+  /**
+   * Rails guards `prune` against re-entry with the `@pruning` flag
+   * (memory_store.rb:110-127): `cleanup` and the per-key deletion below can both
+   * re-enter through a write, and the inner call must not prune again.
+   */
   prune(targetSize: number, maxTime?: number): void {
     if (this.isPruning()) return;
     this._pruning = true;
@@ -196,10 +198,12 @@ export class MemoryStore extends Store implements CacheStore {
 }
 
 export namespace DupCoder {
-  // Mirrors Rails MemoryStore::DupCoder#dump (memory_store.rb:32-38): rebuild the
-  // Entry around a dumped value so the stored value can't alias the caller's,
-  // preserving expires_at/version. Ruby's `entry.value && … != true &&
-  // !Numeric` leaves nil/false/true/Numeric entries untouched.
+  /**
+   * Mirrors Rails MemoryStore::DupCoder#dump (memory_store.rb:32-38): rebuild
+   * the Entry around a dumped value so the stored value can't alias the
+   * caller's, preserving expires_at/version. Ruby's `entry.value && … != true
+   * && !Numeric` leaves nil/false/true/Numeric entries untouched.
+   */
   export function dump(entry: Entry): Entry {
     const value = entry.value;
     if (
@@ -215,13 +219,13 @@ export namespace DupCoder {
     }
   }
 
-  // Mirrors Rails MemoryStore::DupCoder#dump_compressed (memory_store.rb:40-43).
+  /** Mirrors Rails MemoryStore::DupCoder#dump_compressed (memory_store.rb:40-43). */
   export function dumpCompressed(entry: Entry, threshold: number): Entry {
     const compressedEntry = entry.compressed(threshold);
     return compressedEntry.isCompressed() ? compressedEntry : dump(entry);
   }
 
-  // Mirrors Rails MemoryStore::DupCoder#load (memory_store.rb:45-51).
+  /** Mirrors Rails MemoryStore::DupCoder#load (memory_store.rb:45-51). */
   export function load(entry: Entry): Entry {
     if (!entry.isCompressed() && typeof entry.value === "string") {
       return new Entry(loadValue(entry.value), {
@@ -233,17 +237,21 @@ export namespace DupCoder {
     }
   }
 
-  // Rails' `MARSHAL_SIGNATURE` (memory_store.rb:54) is the two leading bytes
-  // every `Marshal.dump` payload carries, which is how `load_value` tells a
-  // serialized value from a String stored verbatim. The trails Marshal
-  // equivalent is the fidelity Coder (coder.ts), whose JSON output carries no
-  // such self-identifying prefix, so dump_value prepends the same signature and
-  // load_value strips it — same discriminator, same two arms.
+  /**
+   * Rails' `MARSHAL_SIGNATURE` (memory_store.rb:54) is the two leading bytes
+   * every `Marshal.dump` payload carries, which is how `load_value` tells a
+   * serialized value from a String stored verbatim. The trails Marshal
+   * equivalent is the fidelity Coder (coder.ts), whose JSON output carries no
+   * such self-identifying prefix, so `dump_value` prepends the same signature
+   * and `load_value` strips it — same discriminator, same two arms.
+   */
   const MARSHAL_SIGNATURE = "\x04\x08";
 
-  // Mirrors Rails MemoryStore::DupCoder#dump_value (memory_store.rb:56-62).
-  // Ruby's `value.dup` guards against the caller mutating the stored String;
-  // JS strings are immutable, so the String arm returns it as-is.
+  /**
+   * Mirrors Rails MemoryStore::DupCoder#dump_value (memory_store.rb:56-62).
+   * Ruby's `value.dup` guards against the caller mutating the stored String;
+   * JS strings are immutable, so the String arm returns it as-is.
+   */
   function dumpValue(value: unknown): string {
     if (typeof value === "string" && !value.startsWith(MARSHAL_SIGNATURE)) {
       return value;
@@ -252,7 +260,7 @@ export namespace DupCoder {
     }
   }
 
-  // Mirrors Rails MemoryStore::DupCoder#load_value (memory_store.rb:64-70).
+  /** Mirrors Rails MemoryStore::DupCoder#load_value (memory_store.rb:64-70). */
   function loadValue(string: string): unknown {
     if (string.startsWith(MARSHAL_SIGNATURE)) {
       return coder.load(string.slice(MARSHAL_SIGNATURE.length));
